Achieved full object capsulation

+++ Cookpot-Backend/cookpot-bl/DataModel/Ingredient.cs
@@ -6,8 +6,8 @@ namespace cookpot.bl.DataModel
     public class Ingredient
     {
         public string Name { get; set; }
-        public Nullable<float> Amount { get; set; }
-        public Nullable<float> Measure { get; set; }
+        public Nullable<double> Amount { get; set; }
+        public Nullable<double> Measure { get; set; }
         public UnitofMeasurement Unit { get; set; }
